refactor(toolkit): pull out interesting parts of cxapp
