+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/cloud-assembly/index.ts
@@ -1,1 +1,4 @@
+export * from './environment';
+export * from './stack-assembly';
+export * from './stack-collection';
 export * from './stack-selector';
