+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/private.ts
@@ -1,0 +1,1 @@
+export * from './io/private';
