+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/private/stack-assembly.ts
@@ -1,6 +1,6 @@
 import type * as cxapi from '@aws-cdk/cx-api';
 import { major } from 'semver';
-import { CloudAssembly, sanitizePatterns, StackCollection, ExtendedStackSelection as CliExtendedStackSelection } from '../../aws-cdk';
+import { BaseStackAssembly, StackCollection, ExtendedStackSelection as CliExtendedStackSelection } from '../../shared-private';
 import { ToolkitError } from '../../shared-public';
 import type { StackSelector } from '../stack-selector';
 import { ExpandStackSelection, StackSelectionStrategy } from '../stack-selector';
@@ -9,7 +9,7 @@ import type { ICloudAssemblySource } from '../types';
 /**
  * A single Cloud Assembly wrapped to provide additional stack operations.
  */
-export class StackAssembly extends CloudAssembly implements ICloudAssemblySource {
+export class StackAssembly extends BaseStackAssembly implements ICloudAssemblySource {
   public async produce(): Promise<cxapi.CloudAssembly> {
     return this.assembly;
   }
@@ -29,7 +29,7 @@ export class StackAssembly extends CloudAssembly implements ICloudAssemblySource
     }
 
     const extend = expandToExtendEnum(selector.expand);
-    const patterns = sanitizePatterns(selector.patterns ?? []);
+    const patterns = StackAssembly.sanitizePatterns(selector.patterns ?? []);
 
     switch (selector.strategy) {
       case StackSelectionStrategy.ALL_STACKS:
