+++ packages/aws-cdk/lib/api/index.ts
@@ -3,3 +3,8 @@ export * from './garbage-collection/garbage-collector';
 export * from './deployments';
 export * from './toolkit-info';
 export * from './aws-auth';
+export * from './rwlock';
+export * from './tree';
+
+export * from '../../../@aws-cdk/tmp-toolkit-helpers/src/api/toolkit-error';
+export * from '../../../@aws-cdk/tmp-toolkit-helpers/src/api/cloud-assembly';
