+++ packages/aws-cdk/lib/cxapp/index.ts
@@ -1,0 +1,4 @@
+export * from './cloud-assembly';
+export * from './cloud-executable';
+export * from './environments';
+export * from './exec';
