+++ packages/aws-cdk/test/cxapp/cloud-executable.test.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/order */
 import * as cxschema from '@aws-cdk/cloud-assembly-schema';
-import { DefaultSelection } from '../../../lib/api/cxapp/cloud-assembly';
-import { registerContextProvider } from '../../../lib/context-providers';
-import { MockCloudExecutable } from '../../_helpers/assembly';
+import { DefaultSelection } from '../../lib/cxapp/cloud-assembly';
+import { registerContextProvider } from '../../lib/context-providers';
+import { MockCloudExecutable } from '../_helpers/assembly';
 
 describe('AWS::CDK::Metadata', () => {
   test('is not generated for new frameworks', async () => {
